Network.AWS.CloudWatchEvents

Description

Amazon CloudWatch Events helps you to respond to state changes in your AWS resources. When your resources change state, they automatically send events into an event stream. You can create rules that match selected events in the stream and route them to targets to take action. You can also use rules to take action on a pre-determined schedule. For example, you can configure rules to:

Automatically invoke an AWS Lambda function to update DNS entries when an event notifies you that Amazon EC2 instance enters the running state.

Direct specific API records from CloudTrail to an Amazon Kinesis stream for detailed analysis of potential security or availability risks.

Periodically invoke a built-in target to create a snapshot of an Amazon EBS volume.

For more information about the features of Amazon CloudWatch Events, see the Amazon CloudWatch Events User Guide .

Service Configuration

cloudWatchEvents :: Service

API version 2015-10-07 of the Amazon CloudWatch Events SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudWatchEvents.

PolicyLengthExceededException

_PolicyLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The event bus policy is too long. For more information, see the limits.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

There is concurrent modification on a rule or target.

InvalidEventPatternException

_InvalidEventPatternException :: AsError a => Getting (First ServiceError) a ServiceError

The event pattern is not valid.

InternalException

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

This exception occurs due to unexpected causes.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

An entity that you specified does not exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You tried to create more rules or add more targets to a rule than is allowed.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

RemoveTargets

module Network.AWS.CloudWatchEvents.RemoveTargets

DeleteRule

module Network.AWS.CloudWatchEvents.DeleteRule

ListRules

module Network.AWS.CloudWatchEvents.ListRules

PutRule

module Network.AWS.CloudWatchEvents.PutRule

DisableRule

module Network.AWS.CloudWatchEvents.DisableRule

PutPermission

module Network.AWS.CloudWatchEvents.PutPermission

ListTargetsByRule

module Network.AWS.CloudWatchEvents.ListTargetsByRule

RemovePermission

module Network.AWS.CloudWatchEvents.RemovePermission

DescribeRule

module Network.AWS.CloudWatchEvents.DescribeRule

EnableRule

module Network.AWS.CloudWatchEvents.EnableRule

ListRuleNamesByTarget

module Network.AWS.CloudWatchEvents.ListRuleNamesByTarget

TestEventPattern

module Network.AWS.CloudWatchEvents.TestEventPattern

DescribeEventBus

module Network.AWS.CloudWatchEvents.DescribeEventBus

PutTargets

module Network.AWS.CloudWatchEvents.PutTargets

PutEvents

module Network.AWS.CloudWatchEvents.PutEvents

Types

RuleState

data RuleState

Constructors

Disabled

Enabled

EcsParameters

data EcsParameters

The custom parameters to be used when the target is an Amazon ECS cluster.

See: ecsParameters smart constructor.

ecsParameters

Arguments

:: Text
epTaskDefinitionARN

-> EcsParameters

Creates a value of EcsParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

epTaskCount - The number of tasks to create based on the TaskDefinition . The default is one.

epTaskDefinitionARN - The ARN of the task definition to use if the event target is an Amazon ECS cluster.

epTaskCount :: Lens' EcsParameters (Maybe Natural)

The number of tasks to create based on the TaskDefinition . The default is one.

epTaskDefinitionARN :: Lens' EcsParameters Text

The ARN of the task definition to use if the event target is an Amazon ECS cluster.

InputTransformer

data InputTransformer

Contains the parameters needed for you to provide custom input to a target based on one or more pieces of data extracted from the event.

See: inputTransformer smart constructor.

inputTransformer

Arguments

:: Text
itInputTemplate

-> InputTransformer

Creates a value of InputTransformer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

itInputPathsMap - Map of JSON paths to be extracted from the event. These are key-value pairs, where each value is a JSON path. You must use JSON dot notation, not bracket notation.

itInputTemplate - Input template where you can use the values of the keys from InputPathsMap to customize the data sent to the target.

itInputPathsMap :: Lens' InputTransformer (HashMap Text Text)

Map of JSON paths to be extracted from the event. These are key-value pairs, where each value is a JSON path. You must use JSON dot notation, not bracket notation.

itInputTemplate :: Lens' InputTransformer Text

Input template where you can use the values of the keys from InputPathsMap to customize the data sent to the target.

KinesisParameters

data KinesisParameters

This object enables you to specify a JSON path to extract from the event and use as the partition key for the Amazon Kinesis stream, so that you can control the shard to which the event goes. If you do not include this parameter, the default is to use the eventId as the partition key.

See: kinesisParameters smart constructor.

kinesisParameters

Arguments

:: Text
kpPartitionKeyPath

-> KinesisParameters

Creates a value of KinesisParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kpPartitionKeyPath - The JSON path to be extracted from the event and used as the partition key. For more information, see Amazon Kinesis Streams Key Concepts in the Amazon Kinesis Streams Developer Guide .

kpPartitionKeyPath :: Lens' KinesisParameters Text

The JSON path to be extracted from the event and used as the partition key. For more information, see Amazon Kinesis Streams Key Concepts in the Amazon Kinesis Streams Developer Guide .

PutEventsRequestEntry

data PutEventsRequestEntry

Represents an event to be submitted.

See: putEventsRequestEntry smart constructor.

putEventsRequestEntry :: PutEventsRequestEntry

Creates a value of PutEventsRequestEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pereTime - The timestamp of the event, per RFC3339 . If no timestamp is provided, the timestamp of the PutEvents call is used.

pereDetailType - Free-form string used to decide what fields to expect in the event detail.

pereResources - AWS resources, identified by Amazon Resource Name (ARN), which the event primarily concerns. Any number, including zero, may be present.

pereSource - The source of the event.

pereDetail - In the JSON sense, an object containing fields, which may also contain nested subobjects. No constraints are imposed on its contents.

pereTime :: Lens' PutEventsRequestEntry (Maybe UTCTime)

The timestamp of the event, per RFC3339 . If no timestamp is provided, the timestamp of the PutEvents call is used.

pereDetailType :: Lens' PutEventsRequestEntry (Maybe Text)

Free-form string used to decide what fields to expect in the event detail.

pereResources :: Lens' PutEventsRequestEntry [Text]

AWS resources, identified by Amazon Resource Name (ARN), which the event primarily concerns. Any number, including zero, may be present.

pereSource :: Lens' PutEventsRequestEntry (Maybe Text)

The source of the event.

pereDetail :: Lens' PutEventsRequestEntry (Maybe Text)

In the JSON sense, an object containing fields, which may also contain nested subobjects. No constraints are imposed on its contents.

PutEventsResultEntry

data PutEventsResultEntry

Represents an event that failed to be submitted.

See: putEventsResultEntry smart constructor.

putEventsResultEntry :: PutEventsResultEntry

Creates a value of PutEventsResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pereErrorCode - The error code that indicates why the event submission failed.

pereErrorMessage - The error message that explains why the event submission failed.

pereEventId - The ID of the event.

pereErrorCode :: Lens' PutEventsResultEntry (Maybe Text)

The error code that indicates why the event submission failed.

pereErrorMessage :: Lens' PutEventsResultEntry (Maybe Text)

The error message that explains why the event submission failed.

pereEventId :: Lens' PutEventsResultEntry (Maybe Text)

The ID of the event.

PutTargetsResultEntry

data PutTargetsResultEntry

Represents a target that failed to be added to a rule.

See: putTargetsResultEntry smart constructor.

putTargetsResultEntry :: PutTargetsResultEntry

Creates a value of PutTargetsResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptreTargetId - The ID of the target.

ptreErrorCode - The error code that indicates why the target addition failed. If the value is ConcurrentModificationException , too many requests were made at the same time.

ptreErrorMessage - The error message that explains why the target addition failed.

ptreTargetId :: Lens' PutTargetsResultEntry (Maybe Text)

The ID of the target.

ptreErrorCode :: Lens' PutTargetsResultEntry (Maybe Text)

The error code that indicates why the target addition failed. If the value is ConcurrentModificationException , too many requests were made at the same time.

ptreErrorMessage :: Lens' PutTargetsResultEntry (Maybe Text)

The error message that explains why the target addition failed.

RemoveTargetsResultEntry

data RemoveTargetsResultEntry

Represents a target that failed to be removed from a rule.

See: removeTargetsResultEntry smart constructor.

removeTargetsResultEntry :: RemoveTargetsResultEntry

Creates a value of RemoveTargetsResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtreTargetId - The ID of the target.

rtreErrorCode - The error code that indicates why the target removal failed. If the value is ConcurrentModificationException , too many requests were made at the same time.

rtreErrorMessage - The error message that explains why the target removal failed.

rtreTargetId :: Lens' RemoveTargetsResultEntry (Maybe Text)

The ID of the target.

rtreErrorCode :: Lens' RemoveTargetsResultEntry (Maybe Text)

The error code that indicates why the target removal failed. If the value is ConcurrentModificationException , too many requests were made at the same time.

rtreErrorMessage :: Lens' RemoveTargetsResultEntry (Maybe Text)

The error message that explains why the target removal failed.

Rule

data Rule

Contains information about a rule in Amazon CloudWatch Events.

See: rule smart constructor.

rule :: Rule

Creates a value of Rule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rEventPattern - The event pattern of the rule. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

rState - The state of the rule.

rARN - The Amazon Resource Name (ARN) of the rule.

rScheduleExpression - The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

rName - The name of the rule.

rDescription - The description of the rule.

rRoleARN - The Amazon Resource Name (ARN) of the role that is used for target invocation.

rEventPattern :: Lens' Rule (Maybe Text)

The event pattern of the rule. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

rState :: Lens' Rule (Maybe RuleState)

The state of the rule.

rARN :: Lens' Rule (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

rScheduleExpression :: Lens' Rule (Maybe Text)

The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

rName :: Lens' Rule (Maybe Text)

The name of the rule.

rDescription :: Lens' Rule (Maybe Text)

The description of the rule.

rRoleARN :: Lens' Rule (Maybe Text)

The Amazon Resource Name (ARN) of the role that is used for target invocation.

RunCommandParameters

data RunCommandParameters

This parameter contains the criteria (either InstanceIds or a tag) used to specify which EC2 instances are to be sent the command.

See: runCommandParameters smart constructor.

runCommandParameters

Arguments

:: NonEmpty RunCommandTarget
rcpRunCommandTargets

-> RunCommandParameters

Creates a value of RunCommandParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcpRunCommandTargets - Currently, we support including only one RunCommandTarget block, which specifies either an array of InstanceIds or a tag.

rcpRunCommandTargets :: Lens' RunCommandParameters (NonEmpty RunCommandTarget)

Currently, we support including only one RunCommandTarget block, which specifies either an array of InstanceIds or a tag.

RunCommandTarget

data RunCommandTarget

Information about the EC2 instances that are to be sent the command, specified as key-value pairs. Each RunCommandTarget block can include only one key, but this key may specify multiple values.

See: runCommandTarget smart constructor.

runCommandTarget

Arguments

:: Text
rctKey

-> NonEmpty Text
rctValues

-> RunCommandTarget

Creates a value of RunCommandTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rctKey - Can be either tag: tag-key or InstanceIds .

rctValues - If Key is tag: tag-key , Values is a list of tag values. If Key is InstanceIds , Values is a list of Amazon EC2 instance IDs.

rctKey :: Lens' RunCommandTarget Text

Can be either tag: tag-key or InstanceIds .

rctValues :: Lens' RunCommandTarget (NonEmpty Text)

If Key is tag: tag-key , Values is a list of tag values. If Key is InstanceIds , Values is a list of Amazon EC2 instance IDs.

Target

data Target

Targets are the resources to be invoked when a rule is triggered. Target types include EC2 instances, AWS Lambda functions, Amazon Kinesis streams, Amazon ECS tasks, AWS Step Functions state machines, Run Command, and built-in targets.

See: target smart constructor.

target

Arguments

:: Text
tId

-> Text
tARN

-> Target

Creates a value of Target with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tRunCommandParameters - Parameters used when you are using the rule to invoke Amazon EC2 Run Command.

tKinesisParameters - The custom parameter you can use to control shard assignment, when the target is an Amazon Kinesis stream. If you do not include this parameter, the default is to use the eventId as the partition key.

tInputTransformer - Settings to enable you to provide custom input to a target based on certain event data. You can extract one or more key-value pairs from the event and then use that data to send customized input to the target.

tInput - Valid JSON text passed to the target. In this case, nothing from the event itself is passed to the target. You must use JSON dot notation, not bracket notation. For more information, see The JavaScript Object Notation (JSON) Data Interchange Format .

tEcsParameters - Contains the Amazon ECS task definition and task count to be used, if the event target is an Amazon ECS task. For more information about Amazon ECS tasks, see Task Definitions in the Amazon EC2 Container Service Developer Guide .

tInputPath - The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. You must use JSON dot notation, not bracket notation. For more information about JSON paths, see JSONPath .

tRoleARN - The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. If one rule triggers multiple targets, you can use a different IAM role for each target.

tId - The ID of the target.

tARN - The Amazon Resource Name (ARN) of the target.

tRunCommandParameters :: Lens' Target (Maybe RunCommandParameters)

Parameters used when you are using the rule to invoke Amazon EC2 Run Command.

tKinesisParameters :: Lens' Target (Maybe KinesisParameters)

The custom parameter you can use to control shard assignment, when the target is an Amazon Kinesis stream. If you do not include this parameter, the default is to use the eventId as the partition key.

tInputTransformer :: Lens' Target (Maybe InputTransformer)

Settings to enable you to provide custom input to a target based on certain event data. You can extract one or more key-value pairs from the event and then use that data to send customized input to the target.

tInput :: Lens' Target (Maybe Text)

Valid JSON text passed to the target. In this case, nothing from the event itself is passed to the target. You must use JSON dot notation, not bracket notation. For more information, see The JavaScript Object Notation (JSON) Data Interchange Format .

tEcsParameters :: Lens' Target (Maybe EcsParameters)

Contains the Amazon ECS task definition and task count to be used, if the event target is an Amazon ECS task. For more information about Amazon ECS tasks, see Task Definitions in the Amazon EC2 Container Service Developer Guide .

tInputPath :: Lens' Target (Maybe Text)

The value of the JSONPath that is used for extracting part of the matched event when passing it to the target. You must use JSON dot notation, not bracket notation. For more information about JSON paths, see JSONPath .

tRoleARN :: Lens' Target (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role to be used for this target when the rule is triggered. If one rule triggers multiple targets, you can use a different IAM role for each target.

tId :: Lens' Target Text

The ID of the target.

tARN :: Lens' Target Text

The Amazon Resource Name (ARN) of the target.