Network.AWS.CloudWatchEvents.DescribeRule

Description

Describes the specified rule.

Creating a Request

describeRule

Arguments

:: Text
desName

-> DescribeRule

Creates a value of DescribeRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desName - The name of the rule.

data DescribeRule

See: describeRule smart constructor.

Request Lenses

desName :: Lens' DescribeRule Text

The name of the rule.

Destructuring the Response

describeRuleResponse

Arguments

:: Int
drrsResponseStatus

-> DescribeRuleResponse

Creates a value of DescribeRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrsEventPattern - The event pattern. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

drrsState - Specifies whether the rule is enabled or disabled.

drrsARN - The Amazon Resource Name (ARN) of the rule.

drrsScheduleExpression - The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

drrsName - The name of the rule.

drrsDescription - The description of the rule.

drrsRoleARN - The Amazon Resource Name (ARN) of the IAM role associated with the rule.

drrsResponseStatus - -- | The response status code.

data DescribeRuleResponse

See: describeRuleResponse smart constructor.

Response Lenses

drrsEventPattern :: Lens' DescribeRuleResponse (Maybe Text)

The event pattern. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

drrsState :: Lens' DescribeRuleResponse (Maybe RuleState)

Specifies whether the rule is enabled or disabled.

drrsARN :: Lens' DescribeRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

drrsScheduleExpression :: Lens' DescribeRuleResponse (Maybe Text)

The scheduling expression. For example, "cron(0 20 * * ? *)", "rate(5 minutes)".

drrsName :: Lens' DescribeRuleResponse (Maybe Text)

The name of the rule.

drrsDescription :: Lens' DescribeRuleResponse (Maybe Text)

The description of the rule.

drrsRoleARN :: Lens' DescribeRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role associated with the rule.

drrsResponseStatus :: Lens' DescribeRuleResponse Int

- | The response status code.