Network.AWS.CloudWatchEvents.ListRuleNamesByTarget

Description

Lists the rules for the specified target. You can see which of the rules in Amazon CloudWatch Events can invoke a specific target in your account.

Creating a Request

listRuleNamesByTarget

Arguments

:: Text
lrnbtTargetARN

-> ListRuleNamesByTarget

Creates a value of ListRuleNamesByTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrnbtNextToken - The token returned by a previous call to retrieve the next set of results.

lrnbtLimit - The maximum number of results to return.

lrnbtTargetARN - The Amazon Resource Name (ARN) of the target resource.

data ListRuleNamesByTarget

See: listRuleNamesByTarget smart constructor.

Request Lenses

lrnbtNextToken :: Lens' ListRuleNamesByTarget (Maybe Text)

The token returned by a previous call to retrieve the next set of results.

lrnbtLimit :: Lens' ListRuleNamesByTarget (Maybe Natural)

The maximum number of results to return.

lrnbtTargetARN :: Lens' ListRuleNamesByTarget Text

The Amazon Resource Name (ARN) of the target resource.

Destructuring the Response

listRuleNamesByTargetResponse

Arguments

:: Int
lrnbtrsResponseStatus

-> ListRuleNamesByTargetResponse

Creates a value of ListRuleNamesByTargetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrnbtrsRuleNames - The names of the rules that can invoke the given target.

lrnbtrsNextToken - Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

lrnbtrsResponseStatus - -- | The response status code.

data ListRuleNamesByTargetResponse

See: listRuleNamesByTargetResponse smart constructor.

Response Lenses

lrnbtrsRuleNames :: Lens' ListRuleNamesByTargetResponse [Text]

The names of the rules that can invoke the given target.

lrnbtrsNextToken :: Lens' ListRuleNamesByTargetResponse (Maybe Text)

Indicates whether there are additional results to retrieve. If there are no more results, the value is null.

lrnbtrsResponseStatus :: Lens' ListRuleNamesByTargetResponse Int

- | The response status code.