Network.AWS.CloudWatchEvents.PutEvents

Description

Sends custom events to Amazon CloudWatch Events so that they can be matched to rules.

Creating a Request

putEvents

Arguments

:: NonEmpty PutEventsRequestEntry
peEntries

-> PutEvents

Creates a value of PutEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

peEntries - The entry that defines an event in your system. You can specify several parameters for the entry such as the source and type of the event, resources associated with the event, and so on.

data PutEvents

See: putEvents smart constructor.

Request Lenses

peEntries :: Lens' PutEvents (NonEmpty PutEventsRequestEntry)

The entry that defines an event in your system. You can specify several parameters for the entry such as the source and type of the event, resources associated with the event, and so on.

Destructuring the Response

putEventsResponse

Arguments

:: Int
persResponseStatus

-> PutEventsResponse

Creates a value of PutEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

persFailedEntryCount - The number of failed entries.

persEntries - The successfully and unsuccessfully ingested events results. If the ingestion was successful, the entry has the event ID in it. Otherwise, you can use the error code and error message to identify the problem with the entry.

persResponseStatus - -- | The response status code.

data PutEventsResponse

See: putEventsResponse smart constructor.

Response Lenses

persFailedEntryCount :: Lens' PutEventsResponse (Maybe Int)

The number of failed entries.

persEntries :: Lens' PutEventsResponse [PutEventsResultEntry]

The successfully and unsuccessfully ingested events results. If the ingestion was successful, the entry has the event ID in it. Otherwise, you can use the error code and error message to identify the problem with the entry.

persResponseStatus :: Lens' PutEventsResponse Int

- | The response status code.