Network.AWS.CloudWatchEvents.PutRule

Description

Creates or updates the specified rule. Rules are enabled by default, or based on value of the state. You can disable a rule using DisableRule .

When you create or update a rule, incoming events might not immediately start matching to new or updated rules. Please allow a short period of time for changes to take effect.

A rule must contain at least an EventPattern or ScheduleExpression. Rules with EventPatterns are triggered when a matching event is observed. Rules with ScheduleExpressions self-trigger based on the given schedule. A rule can have both an EventPattern and a ScheduleExpression, in which case the rule triggers on matching events as well as on a schedule.

Most services in AWS treat : or / as the same character in Amazon Resource Names (ARNs). However, CloudWatch Events uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.

Creating a Request

putRule

Arguments

:: Text
prName

-> PutRule

Creates a value of PutRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prEventPattern - The event pattern. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

prState - Indicates whether the rule is enabled or disabled.

prScheduleExpression - The scheduling expression. For example, "cron(0 20 * * ? *)" or "rate(5 minutes)".

prDescription - A description of the rule.

prRoleARN - The Amazon Resource Name (ARN) of the IAM role associated with the rule.

prName - The name of the rule that you are creating or updating.

data PutRule

See: putRule smart constructor.

Request Lenses

prEventPattern :: Lens' PutRule (Maybe Text)

The event pattern. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

prState :: Lens' PutRule (Maybe RuleState)

Indicates whether the rule is enabled or disabled.

prScheduleExpression :: Lens' PutRule (Maybe Text)

The scheduling expression. For example, "cron(0 20 * * ? *)" or "rate(5 minutes)".

prDescription :: Lens' PutRule (Maybe Text)

A description of the rule.

prRoleARN :: Lens' PutRule (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role associated with the rule.

prName :: Lens' PutRule Text

The name of the rule that you are creating or updating.

Destructuring the Response

putRuleResponse

Arguments

:: Int
prrsResponseStatus

-> PutRuleResponse

Creates a value of PutRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prrsRuleARN - The Amazon Resource Name (ARN) of the rule.

prrsResponseStatus - -- | The response status code.

data PutRuleResponse

See: putRuleResponse smart constructor.

Response Lenses

prrsRuleARN :: Lens' PutRuleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

prrsResponseStatus :: Lens' PutRuleResponse Int

- | The response status code.