Network.AWS.CloudWatchEvents.PutTargets

Description

Adds the specified targets to the specified rule, or updates the targets if they are already associated with the rule.

Targets are the resources that are invoked when a rule is triggered.

You can configure the following as targets for CloudWatch Events:

EC2 instances

AWS Lambda functions

Streams in Amazon Kinesis Streams

Delivery streams in Amazon Kinesis Firehose

Amazon ECS tasks

AWS Step Functions state machines

Amazon SNS topics

Amazon SQS queues

Note that creating rules with built-in targets is supported only in the AWS Management Console.

For some target types, PutTargets provides target-specific parameters. If the target is an Amazon Kinesis stream, you can optionally specify which shard the event goes to by using the KinesisParameters argument. To invoke a command on multiple EC2 instances with one rule, you can use the RunCommandParameters field.

To be able to make API calls against the resources that you own, Amazon CloudWatch Events needs the appropriate permissions. For AWS Lambda and Amazon SNS resources, CloudWatch Events relies on resource-based policies. For EC2 instances, Amazon Kinesis streams, and AWS Step Functions state machines, CloudWatch Events relies on IAM roles that you specify in the RoleARN argument in PutTargets . For more information, see Authentication and Access Control in the Amazon CloudWatch Events User Guide .

If another AWS account is in the same region and has granted you permission (using PutPermission ), you can set that account's event bus as a target of the rules in your account. To send the matched events to the other account, specify that account's event bus as the Arn when you run PutTargets . For more information about enabling cross-account events, see PutPermission .

Input , InputPath and InputTransformer are mutually exclusive and optional parameters of a target. When a rule is triggered due to a matched event:

If none of the following arguments are specified for a target, then the entire event is passed to the target in JSON form (unless the target is Amazon EC2 Run Command or Amazon ECS task, in which case nothing from the event is passed to the target).

If Input is specified in the form of valid JSON, then the matched event is overridden with this constant.

If InputPath is specified in the form of JSONPath (for example, > .detail ), then only the part of the event specified in the path is passed to the target (for example, only the detail part of the event is passed).

If InputTransformer is specified, then one or more specified JSONPaths are extracted from the event and used as values in a template that you specify as the input to the target.

When you specify Input , InputPath , or InputTransformer , you must use JSON dot notation, not bracket notation.

When you add targets to a rule and the associated rule triggers soon after, new or updated targets might not be immediately invoked. Please allow a short period of time for changes to take effect.

This action can partially fail if too many requests are made at the same time. If that happens, FailedEntryCount is non-zero in the response and each entry in FailedEntries provides the ID of the failed target and the error code.

Creating a Request

putTargets

Arguments

:: Text
ptRule

-> NonEmpty Target
ptTargets

-> PutTargets

Creates a value of PutTargets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptRule - The name of the rule.

ptTargets - The targets to update or add to the rule.

data PutTargets

See: putTargets smart constructor.

Request Lenses

ptRule :: Lens' PutTargets Text

The name of the rule.

ptTargets :: Lens' PutTargets (NonEmpty Target)

The targets to update or add to the rule.

Destructuring the Response

putTargetsResponse

Arguments

:: Int
ptrsResponseStatus

-> PutTargetsResponse

Creates a value of PutTargetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptrsFailedEntryCount - The number of failed entries.

ptrsFailedEntries - The failed target entries.

ptrsResponseStatus - -- | The response status code.

data PutTargetsResponse

See: putTargetsResponse smart constructor.

Response Lenses

ptrsFailedEntryCount :: Lens' PutTargetsResponse (Maybe Int)

The number of failed entries.

ptrsFailedEntries :: Lens' PutTargetsResponse [PutTargetsResultEntry]

The failed target entries.

ptrsResponseStatus :: Lens' PutTargetsResponse Int

- | The response status code.