Network.AWS.CloudWatchEvents.TestEventPattern

Description

Tests whether the specified event pattern matches the provided event.

Most services in AWS treat : or / as the same character in Amazon Resource Names (ARNs). However, CloudWatch Events uses an exact match in event patterns and rules. Be sure to use the correct ARN characters when creating event patterns so that they match the ARN syntax in the event you want to match.

Creating a Request

testEventPattern

Arguments

:: Text
tepEventPattern

-> Text
tepEvent

-> TestEventPattern

Creates a value of TestEventPattern with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tepEventPattern - The event pattern. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

tepEvent - The event, in JSON format, to test against the event pattern.

data TestEventPattern

See: testEventPattern smart constructor.

Request Lenses

tepEventPattern :: Lens' TestEventPattern Text

The event pattern. For more information, see Events and Event Patterns in the Amazon CloudWatch Events User Guide .

tepEvent :: Lens' TestEventPattern Text

The event, in JSON format, to test against the event pattern.

Destructuring the Response

testEventPatternResponse

Arguments

:: Int
teprsResponseStatus

-> TestEventPatternResponse

Creates a value of TestEventPatternResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

teprsResult - Indicates whether the event matches the event pattern.

teprsResponseStatus - -- | The response status code.

data TestEventPatternResponse

See: testEventPatternResponse smart constructor.

Response Lenses

teprsResult :: Lens' TestEventPatternResponse (Maybe Bool)

Indicates whether the event matches the event pattern.

teprsResponseStatus :: Lens' TestEventPatternResponse Int

- | The response status code.